+++ tests/test_client.py
@@ -4,17 +4,10 @@ import pytest
 
 from sparc.client import SparcClient
 
-
+# Test creating a default instance
 def test_class(config_file):
     c = SparcClient(connect=False, config_file=config_file)
     assert len(c.module_names) > 0
-
-
-# Config file tests
-def test_config_non_existing(config_file=None):
-    with pytest.raises(RuntimeError):
-        a = SparcClient(config_file, connect=False)
-
 
 # Test config file with incorrect section pointer
 def test_config_no_section(test_resources_dir):
@@ -22,13 +15,29 @@ def test_config_no_section(test_resources_dir):
     with pytest.raises(KeyError):
         a = SparcClient(config_file, connect=False)
 
+# Config non existing config
+def test_config_non_existing(config_file=None):
+    client = SparcClient(config_file, connect=False)
+    c = client.get_config()
+    assert c['global']['default_profile'] == 'default'
+    assert c['default']['pennsieve_profile_name'] == 'pennsieve'
 
+# Test proper config provided
+def test_get_config(test_resources_dir):
+    config_file = os.path.join(test_resources_dir, "config.ini")
+    client = SparcClient(config_file, connect=False)
+    c = client.get_config()
+    print(c.sections())
+    assert c['global']['default_profile'] == 'ci'
+    assert c['ci']['pennsieve_profile_name'] == 'ci'
+
+# Test module addition
 def test_failed_add_module(config_file):
     client = SparcClient(connect=False, config_file=config_file)
     with pytest.raises(ModuleNotFoundError):
         client.add_module(paths="sparc.client.xyz", connect=False)
 
-
+# Test using a config with the module
 def test_add_module_connect(config_file):
     sc = SparcClient(config_file=config_file, connect=False)
 
@@ -46,7 +55,7 @@ def test_add_module_connect(config_file):
     assert d.init_config_arg == expected_module_config
     assert d.connect_method_called is True
 
-
+# Test adding a Pennsieve module
 def test_add_pennsieve(config_file):
     sc = SparcClient(config_file=config_file, connect=False)
     assert "pennsieve" in sc.module_names
@@ -55,8 +64,8 @@ def test_add_pennsieve(config_file):
 
     assert isinstance(sc.pennsieve, PennsieveService)
 
-
-def test_connect(config_file, monkeypatch):
+# Test connection to the module
+def test_module_connect(config_file, monkeypatch):
     sc = SparcClient(config_file=config_file, connect=False)
     mock_connect_results = []